# api/swagger.yaml
definitions:
  http.Error:
    properties:
      message:
        type: string
    required:
    - message
    type: object
  http.Response:
    properties:
      message:
        type: string
      value: {}
    required:
    - message
    type: object
  request.CreateBookCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
  request.UserLoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  request.UserOTPVerifyRequest:
    properties:
      auth_code:
        type: string
      email:
        type: string
    required:
    - auth_code
    - email
    type: object
  request.UserRegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  response.AchievementResponse:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  response.BookCategoryDetailResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  response.BookCategoryResponse:
    properties:
      id:
        type: string
      name:
        type: string
    type: object
  response.JwtToken:
    properties:
      token:
        example: SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
        type: string
    type: object
  response.UserAchievementResponse:
    properties:
      achievement:
        items:
          $ref: '#/definitions/response.AchievementResponse'
        type: array
      id:
        type: string
      user_id:
        type: string
    type: object
  response.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
      profession:
        type: string
      user_achievement:
        items:
          $ref: '#/definitions/response.UserAchievementResponse'
        type: array
      username:
        type: string
    type: object
externalDocs:
  description: OpenAPI
  url: https://swagger.io/resources/open-api/
info:
  contact:
    email: kevinnaserwan@gmail.com
    name: Coursphere API
paths:
  /book-category/{ID}:
    get:
      consumes:
      - application/json
      description: Get Book Category By ID Endpoint
      parameters:
      - description: Book Category ID
        in: path
        name: ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  $ref: '#/definitions/response.BookCategoryDetailResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Get Book Category By ID
      tags:
      - Book Category
  /book-category/all:
    get:
      consumes:
      - application/json
      description: Get All Book Categories Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  $ref: '#/definitions/response.BookCategoryResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Get All Book Categories
      tags:
      - Book Category
  /book-category/create:
    post:
      consumes:
      - application/json
      description: Create Book Category Endpoint
      parameters:
      - description: Create Book Category Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateBookCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  $ref: '#/definitions/response.JwtToken'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Create Book Category
      tags:
      - Book Category
  /user/{id}:
    delete:
      consumes:
      - application/json
      description: Delete User Endpoint
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Delete User
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Get User By ID Endpoint
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Get User By ID
      tags:
      - User
  /user/all:
    get:
      consumes:
      - application/json
      description: Get All Users Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  items:
                    $ref: '#/definitions/response.UserResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Get All Users
      tags:
      - User
  /user/create:
    post:
      consumes:
      - application/json
      description: Create User Endpoint
      parameters:
      - description: Create User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserRegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  $ref: '#/definitions/response.JwtToken'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Create User
      tags:
      - User
  /user/login:
    post:
      consumes:
      - application/json
      description: Login User Endpoint
      parameters:
      - description: Login User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserLoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  $ref: '#/definitions/response.JwtToken'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Login User
      tags:
      - User
  /user/resend-otp:
    post:
      consumes:
      - application/json
      description: Resend Verification Email Endpoint
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Resend Verification Email
      tags:
      - User
  /user/verify-otp:
    post:
      consumes:
      - application/json
      description: Verify User OTP Endpoint
      parameters:
      - description: Verify User OTP Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UserOTPVerifyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/http.Response'
            - properties:
                value:
                  $ref: '#/definitions/response.UserResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.Error'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.Error'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.Error'
      summary: Verify User OTP
      tags:
      - User
securityDefinitions:
  AdminAuthorization:
    description: "Admin\tJwt Token Authorization"
    in: header
    name: Authorization
    type: apiKey
  UserAuthorization:
    description: "User \tJwt Token Authorization"
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"